// proj_monitoramento/backend/src/modules/user/Interfaces/ICreateUserAdminDTO.ts
interface ICreateUserAdminDTO {
 
  nome_admin: string; 
  nome_orgao: string;
  cargo: string;
  profissao: string;
  nivel_escolaridade: string;
  sexo: string;
  genero: string;
  cpf_admin: string; 
  mat_orgao: string;
  unidade_sei: string;
  email_institucional: string;
  email_pessoal: string;
  nome_empresa: string;
  num_contrato: string;
  password: string;

  }
  
  export { ICreateUserAdminDTO };
  